`Update.Version` should be an `int64`.

+++ src/mozilla.org/simplepush/client/packet.go
@@ -204,5 +204,5 @@ func (a *ClientACK) MarshalJSON() ([]byte, error) {
 
 type Update struct {
 	ChannelId string `json:"channelID"`
-	Version   int    `json:"version"`
+	Version   int64  `json:"version"`
 }
